Setting up the application to use React Navigation.
