+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "ColorMatch",
-    "slug": "ColorMatch",
+    "name": "OutfitMatch",
+    "slug": "OutfitMatch",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
