+++ package.json
@@ -9,13 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/metro-runtime": "~3.2.1",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.10.1",
     "expo": "~51.0.21",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-native": "0.74.3",
     "react-dom": "18.2.0",
+    "react-native": "0.74.3",
     "react-native-web": "~0.19.10",
-    "@expo/metro-runtime": "~3.2.1"
+    "react-native-screens": "3.31.1",
+    "react-native-safe-area-context": "4.10.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
