+++ screens/HomeScreen.js
@@ -1,0 +1,22 @@
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
+
+export const HomeScreen = ({ navigation }) => {
+
+    return (
+        <View style={styles.container}>
+            <Text>Open up App.js to start working on your app!</Text>
+
+            <StatusBar style="auto" />
+        </View>
+    );
+};
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+});
